% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmp.R
\name{ccmp_dictionary}
\alias{ccmp_dictionary}
\title{Printable dictionary of CCMP variables}
\usage{
ccmp_dictionary()
}
\value{
a data frame of class \code{datamatch_dictionary}
}
\description{
Printable dictionary of CCMP variables
}
\examples{
ccmp_dictionary()
}
\seealso{
\code{\link[=ccmp_variables]{ccmp_variables()}}
}
