% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefi.R
\name{cefi_dictionary}
\alias{cefi_dictionary}
\title{Printable dictionary of CEFI variables}
\usage{
cefi_dictionary()
}
\value{
a data frame of class \code{datamatch_dictionary}
}
\description{
Printable dictionary of CEFI variables
}
\examples{
cefi_dictionary()
}
\seealso{
\code{\link[=cefi_variables]{cefi_variables()}}
}
