% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cefi.R
\name{cefi_reference}
\alias{cefi_reference}
\title{The CEFI citation, which is one reference for every run}
\usage{
cefi_reference()
}
\value{
\if{html}{\out{<char>}} the reference
}
\description{
The CEFI citation, which is one reference for every run
}
\keyword{internal}
