% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retime.R
\name{check_columns}
\alias{check_columns}
\title{Check that requested columns exist}
\usage{
check_columns(env_dat, vars)
}
\arguments{
\item{env_dat}{the source object}

\item{vars}{requested columns}
}
\value{
\code{NULL}, invisibly; called for the error
}
\description{
Check that requested columns exist
}
\keyword{internal}
