% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{erddap_dictionary}
\alias{erddap_dictionary}
\title{Printable dictionary of the ERDDAP datasets and their variables}
\usage{
erddap_dictionary()
}
\value{
a data frame of class \code{datamatch_dictionary}
}
\description{
Printable dictionary of the ERDDAP datasets and their variables
}
\examples{
erddap_dictionary()
}
\seealso{
\code{\link[=erddap_datasets]{erddap_datasets()}}
}
