% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvcom.R
\name{fvcom_dictionary}
\alias{fvcom_dictionary}
\title{Printable dictionary of FVCOM variables}
\usage{
fvcom_dictionary(mesh = c("all", "node", "element"))
}
\arguments{
\item{mesh}{filter to \code{"node"}, \code{"element"}, or \code{"all"}}
}
\value{
a data frame of class \code{datamatch_dictionary}
}
\description{
Printable dictionary of FVCOM variables
}
\examples{
fvcom_dictionary()
fvcom_dictionary("element")
}
\seealso{
\code{\link[=fvcom_variables]{fvcom_variables()}}
}
