% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvcom.R
\name{fvcom_reference}
\alias{fvcom_reference}
\title{The FVCOM citation, which is one paper for every run}
\usage{
fvcom_reference()
}
\value{
\if{html}{\out{<char>}} the reference
}
\description{
The FVCOM citation, which is one paper for every run
}
\keyword{internal}
