% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hycom.R
\name{hycom_dictionary}
\alias{hycom_dictionary}
\title{Printable dictionary of HYCOM variables}
\usage{
hycom_dictionary()
}
\value{
a data frame of class \code{datamatch_dictionary}
}
\description{
Printable dictionary of HYCOM variables
}
\examples{
hycom_dictionary()
}
\seealso{
\code{\link[=hycom_variables]{hycom_variables()}}
}
