% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_indices.R
\name{index_cache_age}
\alias{index_cache_age}
\title{Age of a cached file, in days}
\usage{
index_cache_age(file)
}
\arguments{
\item{file}{path to a cached file}
}
\value{
age in days, or \code{Inf} if it does not exist
}
\description{
Age of a cached file, in days
}
\keyword{internal}
