% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obdaac.R
\name{obdaac_dictionary}
\alias{obdaac_dictionary}
\title{Printable dictionary of OB.DAAC variables}
\usage{
obdaac_dictionary()
}
\value{
a data frame of class \code{datamatch_dictionary}
}
\description{
Printable dictionary of OB.DAAC variables
}
\examples{
obdaac_dictionary()
}
\seealso{
\code{\link[=obdaac_variables]{obdaac_variables()}}, \code{\link[=obdaac_sensors]{obdaac_sensors()}}
}
