% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pretty_steps}
\alias{pretty_steps}
\title{Axis positions that do not overprint}
\usage{
pretty_steps(n)
}
\arguments{
\item{n}{number of steps}
}
\value{
integer positions to label
}
\description{
Axis positions that do not overprint
}
\keyword{internal}
