% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{select_time_step}
\alias{select_time_step}
\title{Resolve a \code{time} argument to a set of rows}
\usage{
select_time_step(env_dat, time)
}
\arguments{
\item{env_dat}{an \code{sf} POINT object}

\item{time}{an index, or a named vector of YEAR/MONTH/DAY values}
}
\value{
a list with \code{rows} and a \code{label}
}
\description{
Resolve a \code{time} argument to a set of rows
}
\keyword{internal}
