% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmp.R
\name{to_180}
\alias{to_180}
\title{Convert a longitude back to the -180 to 180 convention}
\usage{
to_180(x)
}
\arguments{
\item{x}{\if{html}{\out{<numeric>}} longitudes in 0-360}
}
\value{
\if{html}{\out{<numeric>}} the same longitudes, negative west
}
\description{
Convert a longitude back to the -180 to 180 convention
}
\keyword{internal}
