% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessEnvDat.R
\name{accessEnvDat}
\alias{accessEnvDat}
\title{Access environmental data from Copernicus Marine Service}
\usage{
accessEnvDat(
  product_id,
  dataset_id,
  vars,
  years,
  months,
  bounding_box,
  depth = c(0, 1),
  overwrite = FALSE
)
}
\arguments{
\item{product_id}{\if{html}{\out{<char>}} product identification string from Copernicus Marine Data Store}

\item{dataset_id}{\if{html}{\out{<char>}} dataset identification string from Copernicus Marine Data Store}

\item{vars}{\if{html}{\out{<char>}} list of variables to access, can be found under data access tab on Copernicus Marine Data Store,
accepts variable abbreviations only}

\item{years}{\if{html}{\out{<numeric>}} years of data to access}

\item{months}{\if{html}{\out{<numeric>}} months of data to access}

\item{bounding_box}{\if{html}{\out{<list>}} named list of spatial coordinates of bounding box}

\item{depth}{\if{html}{\out{<numeric>}} depth range to access (in meters)}

\item{overwrite}{\if{html}{\out{<logical>}} whether or not to overwrite the data if it exists locally}
}
\value{
envDat \if{html}{\out{<sf object>}} sf object containing requested environmental data from Copernicus Marine Service
}
\description{
Access environmental data from Copernicus Marine Service
}
