% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_indices.R
\name{index_dictionary}
\alias{index_dictionary}
\alias{print.datamatch_index_dictionary}
\title{Printable dictionary of climate indices}
\usage{
index_dictionary()

\method{print}{datamatch_index_dictionary}(x, ...)
}
\arguments{
\item{x}{a \code{datamatch_index_dictionary}}

\item{...}{ignored}
}
\value{
a data frame of class \code{datamatch_index_dictionary}
}
\description{
Printable dictionary of climate indices
}
\examples{
index_dictionary()
}
