% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchData.R
\name{matchData}
\alias{matchData}
\title{Match environmental data to species occurrence data}
\usage{
matchData(speciesDat, envDat)
}
\arguments{
\item{speciesDat}{\if{html}{\out{<sf object>}} species observation data (e.g., presence, density, count);
must have spatial and temporal components}

\item{envDat}{\if{html}{\out{<sf object>}} environmental data accessed using the datamatch::accessEnvDat
function to be matched to the species observation data}
}
\value{
envDat \if{html}{\out{<sf object>}} sf object containing requested environmental data from Copernicus Marine Service
}
\description{
Match environmental data to species occurrence data
}
